// src/interfaces/CategoriaComprobante.ts
export interface CategoriaComprobanteType {
    data:       CategoriaComprobanteData[];
    pagination: ComprobantePagination;
}

export interface CategoriaComprobanteData {
    id:          string;
    descripcion: string;
    id_tipo:     string;
    id_estado:   string;
    id_usuario:  string;
    created_at:  string;
    updated_at:  string;
    deleted_at:  null;
    users:       ComprobanteUsers;
}

export interface ComprobanteUsers {
    id:                string;
    name:              string;
    email:             string;
    email_verified_at: null;
    password:          string;
    remember_token:    null;
    created_at:        null;
    updated_at:        null;
}


export interface ComprobantePagination {
    page:  number;
    limit: number;
    total: number;
    pages: number;
}